Split flavour.rs into 4 files
